# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "REST API for managing tasks with prioritization based on deadline and importance.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List all tasks",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter tasks by status (TODO, IN_PROGRESS, COMPLETED)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a specific task by ID",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted"
                    }
                }
            }
        },
        "/api/tasks/priority": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get prioritized tasks",
                "operationId": "722dbea74b79547da045139784b5a4f5",
                "responses": {
                    "200": {
                        "description": "List of prioritized tasks"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "required": [
                    "title",
                    "status",
                    "importance",
                    "deadline"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Prepare report"
                    },
                    "description": {
                        "type": "string",
                        "example": "Quarterly summary"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "TODO",
                            "IN_PROGRESS",
                            "COMPLETED"
                        ],
                        "example": "TODO"
                    },
                    "importance": {
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 4
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-22 18:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Task Management Endpoints"
        }
    ]
}